requirements.txt
PyYAML
python-dotenv
requests
